+++ src/parser/mod.rs
@@ -4,4 +4,4 @@ pub mod core_parser;
 // 以下はcore_parser traitを実装している
 // core_parser.parser for *
 pub mod expr_parser;
-pub mod state_parser;
+pub mod stmt_parser;
