+++ src/token/block.rs
@@ -1,7 +1,7 @@
 use crate::abs::ast::*;
 
 use crate::errors::parser_errors::ParserError;
-use crate::parser::state_parser::*;
+use crate::parser::stmt_parser::*;
 
 /// # BlockBranch
 /// ブロックを格納するデータstruct
